requirements.txt
# local package
-e .

# external requirements
python-dotenv>=0.5.1
python >=2.6
scikit-learn
pandas
numpy
matplotlib.pyplot
chefboost
docutils
seaborn
scipy
tensorflow
(other packages o mid-level dependencies (if any) can be installed by checking condense.py)

R packages:
pclag
bnlearn